Metadata API: Move signature verification to Key
This is likely not needed by users of the API (as they are interested
in the higher level "verify with threshold of signatures").

Moving this to Key makes the API cleaner because including both
"verify myself" and "verify a delegate with threshold" can look awkward
in Metadata.

sign() and verify_signature() are now in different places which is
unbalanced but explained by the fact that they are done by entirely
different components: signing is done by Signer (no Key needed),
verification is done by Key (no Signer needed).

Signed-off-by: Jussi Kukkonen <jkukkonen@vmware.com>