Metadata API: Move "typed constructors" to Signed
Change the way Metadata[Root] and other typed Metadata are constructed.
Before:
    root_md = Metadata._from_file(filename, signed_type=Root)
after:
    root_md = Root.metadata_from_file(filename)

This nicely removes the requirement for an extra argument.

This unfortunately adds a lot of lines to metadata.py but actual LOC
count does not really change -- it's just 2 additional abstract methods
and 4 one-liner implementations for each of those. The only things
these implementations do are:
 * annotates returns type as e.g. Metadata[Root]
 * raises if the signed type in input is incorrect

Signed-off-by: Jussi Kukkonen <jkukkonen@vmware.com>